requirements.txt
fastapi
slowapi
pydantic_settings
uvicorn
python-multipart
SQLAlchemy
pandas
openpyxl
python-dotenv
